// source/feeds/models/FeedsResult.ts
export interface FeedResult {
  id: number;
  publishDate: Date;
  title: string;
  summary: string;
  content: string;
  authors: string;
  categories: string;
  link: string;
  images: string;
  sourceId: number;
  sourceType: number;
  sourceIcon: string;
  sourceName: string;
  sourceTitle: string;
  sourceDescription: string;
}